extract saveStudents into a file and load it in both envs
